kaf012: update backend

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.dom/src/main/java/nts/uk/ctx/at/request/dom/application/timeleaveapplication/output/TimeVacationRemainingOutput.java
@@ -20,13 +20,13 @@ public class TimeVacationRemainingOutput {
     private int super60HRemainingTime = 0;
 
     // 介護の残日数
-    private int careRemainingDays = 0;
+    private double careRemainingDays = 0;
 
     // 介護の残時間
     private int careRemainingTime = 0;
 
     // 子看護の残日数
-    private int childCareRemainingDays = 0;
+    private double childCareRemainingDays = 0;
 
     // 子看護の残時間
     private int childCareRemainingTime = 0;
